requirements.txt
cassandra-driver
aiohttp
aiocassandra
marshmallow
pynacl
passlib
cryptography
kafka-python
aiokafka
